extras/qr-generator/pyproject.toml
[tool.poetry]
name = "qr-generator"
version = "0.1.0"
description = ""
authors = ["Kaj Siebert <kaj@k-si.com>"]

[tool.poetry.dependencies]
python = "^3.8"
qrcode = "^7.3.1"
Pillow = "^8.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
